Remove code duplecate introduced by PR #156
